+++ packages/domain-clients-core/src/solana/config.ts
@@ -1,0 +1,6 @@
+import type { SolanaCluster, SolanaUrlOrMoniker } from '@/solana';
+
+export interface SolanaConfig {
+  clusters: SolanaCluster[];
+  defaultUrlOrMoniker: SolanaUrlOrMoniker;
+}