+++ packages/domain-clients-core/src/solana/index.ts
@@ -2,3 +2,4 @@ export * from './SolanaClient';
 export * from './BaseClient';
 export * from './SigningSolanaClient';
 export * from './types';
+export * from './config';