+++ packages/domain-clients-react/src/evm/hooks/useSigningEvmClient.tsx
@@ -1,33 +1,29 @@
 'use client';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { useEvmConfig } from '@/evm';
 import { SigningEvmClient } from '@valence-protocol/domain-clients-core/evm';
 import { useAccount, useWalletClient } from 'wagmi';
 
-export function useSigningEvmClient(chainId: number) {
+export function useSigningEvmClient(
+  chainId: number
+): SigningEvmClient | undefined {
   const config = useEvmConfig();
   const account = useAccount();
   const { data: walletClient } = useWalletClient({ chainId });
 
-  const [client, setClient] = useState<SigningEvmClient | undefined>(undefined);
-
-  useEffect(() => {
-    if (!account || !walletClient) {
-      setClient(undefined);
+  return useMemo(() => {
+    if (!account) {
+      console.warn('No evm account found');
       return;
     }
-    const client = new SigningEvmClient({
+    if (!walletClient) {
+      console.warn('No evm wallet client found for chainId: ', chainId);
+      return;
+    }
+    return new SigningEvmClient({
       config: config.wagmiConfig,
       chainId: chainId,
       signer: walletClient,
     });
-    setClient(client);
   }, [account, walletClient, config.wagmiConfig, chainId]);
-
-  return useMemo(
-    () => ({
-      client,
-    }),
-    [client]
-  );
 }