+++ packages/domain-clients-react/src/solana/context/SolanaClientProvider.tsx
@@ -1,0 +1,11 @@
+import { WalletUi } from '@wallet-ui/react';
+import { useSolanaConfig } from '@/solana/hooks';
+
+export const SolanaClientProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const config = useSolanaConfig();
+  return <WalletUi config={config}>{children}</WalletUi>;
+};