+++ packages/domain-clients-react/src/solana/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './useSolanaConfig';
+export * from './useSolanaClient';
+export * from './useSolanaSigningClient';